pyproject.toml
[project]
name = "auto-fb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
    "playwright>=1.58.0",
    "playwright-stealth>=2.0.2",
    "psycopg[binary]>=3.3.3",
    "python-dotenv>=1.2.2",
    "python-ghost-cursor>=0.1.1",
    "sqlalchemy>=2.0.48",
]
